added sync_user_repos test
